split up sources a bit more. also added File.* (just 'read' rn)
